Mixed Q1Net perceptual / MSE loss implementation
