+++ models/__init__.py
@@ -4,3 +4,4 @@ from .rrdbnet import RRDBNet
 from .prism import PrismLumaS4
 from .flare import FlareNet, FlareLuma, FlareChroma
 from .perceptual import Q1Net
+from .loss import MixedQ1MSELoss, Q1PerceptualLoss
